Updated to latest generation

+++ backend/src/main/resources/ddl.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS products (
   id        SERIAL PRIMARY KEY,
   name      VARCHAR(40) NOT NULL,
-  stock     BIGINT
+  stock     INT
 );
 
 DELETE FROM products;
